Add Ada support and improve scripts
Added support for Ada back in by reactivating and updating
install-gnat.py Added Ada build-prove workflow

Refactored venv activation to paths.sh

Added ensure_gnat, which will check for gnat on the path (via which) or
will add a local install of GNAT CE to the path (if it's there) or will
offer to install GNAT CE via install-gnat.py This means the user can
always choose a different GNAT and it will be used; otherwise, the local
install of GNAT CE will be used.

Added wrapper shell scripts for the python testing and proof scripts so
that the environment can be automatically configured.

Updated the C++ workflow

Note: sourcing anod and then running anod as a shell function isn't
behaving quite right in either bash or zsh (and for different reasons).
Needs further investigation.