Enhance run-example to better search for binaries.
run-example now follows similar logic to run-tests in searching for
binaries. uxas and uxas-ada are also special-cased so that searching
locally and in anod should succeed as expected. The search order is:

1. path (using `which`)
2. local (obj/cpp/uxas or src/ada/uxas-ada)
3. anod (inside uxas-release or uxas-ada-release)

run-example also follows similar logic in search for OpenAMASE:

1. specified via argument
2. local (develop/OpenAMASE)
3. anod (inside amse)

I believe these have all been tested.

Help / error strings have been updated so that they are consistent with
the new infrastructure.

Debug output has been improved to be more consistent with that provided
by e3 (but doesn't actually directly use e3, still).

This commit also improves the .gitignore for ada support and updates the
Makefile so that it removes the binary on clean.